Gave up trying to put it all in one file.
